// src/static-client/src/types/FieldNames.ts
// `FieldNames` is an enum that is used to define the names of the form fields.
export enum FieldNames {
    SKOGTYPE = "Skogtype",
    VOLUM_PR_DEKAR = "Volum pr dekar",
    TOMMERTREAR_PR_DEKAR = "Treantall pr dekar",
    ANTALL_DEKAR = "Antall dekar",
    RYDDETREAR_PR_DEKAR = "Ryddetrær pr dekar",
    OVERFLATESTRUKTUR_TERRENG = "Overflatestruktur hogstfelt",
    OVERFLATESTRUKTUR_TRAKTORVEG = "Overflatestruktur basvei",
    HELLING_HOGSTFELT = "Helling kjøring hogstfelt",
    HELLING_PAA_TRAKTORVEG = "Helling kjøring basvei",
    KJOREAVSTAND_TERRENG = "Lengde kjøring hogstfelt",
    KJOREAVSTAND_VEG = "Lengde basvei",
    LASSTORRELSE = "Lasstørrelse",
    ANTALL_SORTIMENT = "Antall sortiment",
    TIMEKOST_HOGSTMASKIN = "Timekostnad - hogstmaskin",
    TIMEKOST_LASSBEARER = "Timekostnad - lassbærer",
    OPPSTARTSKOSTNADER = "Oppstartskostnader",
    ENHETSPRIS_MASKINFLYTT = "Enhetspris pr maskinflytt",
    ANTALL_MASKINFLYTT = "Antall maskinflytt",
    ENHETSPRIS_BRO = "Enhetspris pr bro",
    ANTALL_BRUER = "Midlertidige broer",
    ENHETSPRIS_KLOPP = "Enhetspris pr kloppmeter",
    KLOPPLEGGING = "klopplegging",
    TIMEKOST_GRAVEMASKIN = "Timekostnad - gravemaskin",
    TIMER_GRAVEMASKIN = "Timer gravemaskin",
    MANUELT_TILLEGGSARBEID = "Timekostnad - tilleggsarbeid",
    TIMER_TILLEGGSARBEID = "Timer tilleggsarbeid",


}